♻️ refactor: 컨텐츠 데이터 저장 메서드명을 직관적으로 수정, 추천내실 dto의 패키지 명시가 잘못된 부분 수정
