+++ src/main/java/com/teamProject/lostArkProject/collectible/dao/CollectibleDAO.java
@@ -2,9 +2,8 @@ package com.teamProject.lostArkProject.collectible.dao;
 
 import com.teamProject.lostArkProject.collectible.domain.CollectiblePoint;
 import com.teamProject.lostArkProject.collectible.domain.RecommendCollectible;
-import com.teamProject.lostArkProject.collectible.domain.RecommendCollectibleFullDTO;
 import com.teamProject.lostArkProject.collectible.dto.CollectiblePointSummaryDTO;
-import com.teamProject.lostArkProject.collectible.dto.RecommendCollectibleDetailDTO;
+import com.teamProject.lostArkProject.collectible.dto.RecommendCollectibleFullDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
