+++ src/main/java/com/teamProject/lostArkProject/common/config/ResetDatabaseEvent.java
@@ -17,6 +17,6 @@ public class ResetDatabaseEvent implements ApplicationListener<ApplicationReadyE
     @Override
     public void onApplicationEvent(ApplicationReadyEvent event) {
         log.info("데이터베이스 초기화 작업을 시작합니다.");
-        contentService.saveContent();
+        contentService.saveContents();
     }
 }
